固定类型的字节转换完成

+++ src/main/java/org/gongxuanzhang/mysql/entity/page/PageHeader.java
@@ -31,57 +31,57 @@ public class PageHeader implements ByteSwappable<PageHeader> {
      * 页中slot的数量，2字节
      * MySQL中叫 PAGE_N_DIR_SLOTS
      **/
-    int slotCount;
+    short slotCount;
 
     /**
      * 页中还没使用的最小地址，2字节
      **/
-    int heapTop;
+    short heapTop;
     /**
      * 页中的记录数量 2字节(包括最小记录和最大记录)
      * MySQL中叫 PAGE_N_HEAP
      **/
-    int absoluteRecordCount;
+    short absoluteRecordCount;
     /**
      * 2字节|该页中记录的数量 (不包括最小和最大记录以及被标记为删除的记录)
      **/
-    int recordCount;
+    short recordCount;
     /**
      * 第一个被标记删除的地址，可以通过next_record找到删除列表
      * 2字节
      **/
-    int free;
+    short free;
     /**
      * 2字节|已删除记录占用的字节数
      **/
-    int garbage;
+    short garbage;
     /**
      * 2字节/最后插入记录的位置
      **/
-    int lastInsert;
+    short lastInsert;
     /**
      * 2字节/记录插入的方向
      **/
-    int direction;
+    short direction;
     /**
      * 2字节/一个方向连续插入的记录数量
      **/
-    int directionCount;
+    short directionCount;
 
     /**
      * 8字节/修改当前页的最大事务ID，该值仅在二级索引中定义
      **/
-    int maxTransactionId;
+    long maxTransactionId;
 
     /**
      * 2字节 当前页在B+树中所处的层级
      **/
-    int level;
+    short level;
 
     /**
      * 8字节 索引ID,表示当前页属于哪个索引
      **/
-    int indexId;
+    long indexId;
     /**
      * 10字节 b+树叶子段的头部信息 只有root才有意义
      **/
@@ -89,7 +89,7 @@ public class PageHeader implements ByteSwappable<PageHeader> {
     /**
      * 10字节 b+树 非叶子结点段的头部信息 只有root才有意义
      **/
-    int segTop;
+    long segTop;
 
 
     @Override
